Fix test regressions due to floating point
Not sure if the game supports floating point numbers, but 150% and 200%
scaling end up with whole integers, so maintain that!

Regressed from 53c72926163d17d199b4ee318120fa8768edc61b

+++ patches.py
@@ -128,7 +128,7 @@ def _upscale_uiscript(entry: dbpf.Entry):
             new_values = []
             values = part.split("(")[1].split(")")[0]
             for number in values.split(","):
-                new_values.append(str(int(number) * UI_MULTIPLIER))
+                new_values.append(str(int(int(number) * UI_MULTIPLIER)))
             part = f"{name}={part.replace(values, ','.join(new_values))}"
             output.append(part)
         return "".join(output)
@@ -204,7 +204,7 @@ def upscale_fontstyle_ini(file: GameFile, write_meta_file=True):
             continue
 
         old_size = int(parts[3])
-        new_size = old_size * UI_MULTIPLIER
+        new_size = int(old_size * UI_MULTIPLIER)
         parts[3] = str(new_size)
         output.append('"'.join(parts))
